Create EDA Submission Cohort_Osaka Vaibhav_Lawande
aim of this project was to perform exploratory data analysis (EDA) on a hotel booking dataset in order to gain insights into customer behavior and booking patterns. This data set contains booking information for a city hotel and a resort hotel and includes information such as when the booking was made, length of stay, the number of adults, children, and/or babies, and the number of available parking spaces, among other things.First, we data processing and data manipulation which involved removing missing values, handling duplicates.Then we do Data visualization using various  graphical techniques, and we discovered several intresting insights from this.